+++ Tests/TestPackage/Runtime/Constants.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Unity.XR.Management.TestPackage
-{
-    internal struct Constants
-    {
-        public const string k_SettingsKey = "TestPackageSettings";
-    }
-}